requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==2.4.1
attrs==25.3.0
babel==2.17.0
bitarray==3.3.2
build==1.2.2.post1
cached-property==2.0.1
cbor2==5.6.5
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.1.8
colorama==0.4.6
coverage==7.8.0
cytoolz==1.0.1
dotenv==0.9.9
eth-account==0.13.7
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-stdlib==0.2.8
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.7.0
ghp-import==2.1.0
hexbytes==1.3.0
hypothesis==6.131.10
idna==3.10
immutables==0.21
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
lark==1.2.2
lru-dict==1.3.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.41
mkdocs-material-extensions==1.3.1
multidict==6.5.0
packaging==23.2
paginate==0.5.7
parsimonious==0.10.0
pathspec==0.12.1
pip-tools==7.4.1
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.2
py-ecc==8.0.0
py-evm==0.12.0b3
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pymdown-extensions==10.15
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyunormalize==16.0.0
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rlp==4.1.0
ruff==0.12.0
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
titanoboa==0.2.7
toolz==1.0.0
trie==3.1.0
types-requests==2.32.4.20250611
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
vvm==0.3.2
vyper==0.4.3
watchdog==6.0.0
web3==7.12.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
zipp==3.21.0
